Refactor: 개별 크롤러가 BaseCrawler를 상속받도록 수정
