+++ extract/news/sbs_crawler.py
@@ -1,30 +1,23 @@
-from io import BytesIO
+from datetime import datetime
+from typing import List
 
-import pandas as pd
 import requests
 from bs4 import BeautifulSoup
-from datetime import datetime
-import sys
-import pyarrow as pa
-import pyarrow.parquet as pq
-from conf import BUCKET_NAME, MARKER_PATH
-from type.news_crawler import NewsRequest, NewsResponse
-from typing import List, Optional
-import io
-import boto3
-import json
 
-class SBSCrawler:
-    NEWS_TAG = 'SBS'
+from base_crawler import BaseCrawler
+from type.news_crawler import NewsRequest, NewsResponse
+
+
+class SBSCrawler(BaseCrawler):
     def __init__(self, request : NewsRequest):
-        self.request = request
+        super().__init__(request, source="SBS")
         self.page_base_url = (
             "https://searchapi.news.sbs.co.kr/search/news?" +
             "searchField=all&sectionCd=01%2C02%2C03%2C07%2C08%2C09%2C14&collection=news_sbs&limit=10"
         )
         self.news_base_url = "https://news.sbs.co.kr/news/endPage.do?plink=ARTICLE&cooper=SBSNEWSSEARCH"
 
-    def get_search_result(self)-> Optional[List[NewsResponse]]:
+    def get_search_result(self) -> List[NewsResponse]:
         # 페이지 선택
         news_input = self.request
         news_output_total = []
@@ -78,78 +71,9 @@ class SBSCrawler:
                             post_time=news_date,
                             title=news_title,
                             content=news_body,
-                            source=self.NEWS_TAG,
+                            source=self.source,
                             link=news_url,
                             keyword=self.request["keyword"]
                         )
                     )
         return news_output
-
-    def save_to_parquet(self, news_output:List[NewsResponse], f_name:str)->None:
-        if len(news_output) != 0:
-            df = pd.DataFrame(news_output, columns=["post_time", "title", "content", "source", "link","keyword"])
-            df.to_parquet(f_name, engine="pyarrow")
-        else:
-            print("news_output is empty", file=sys.stderr)
-    
-    def upload_s3(self, news_output:List[NewsResponse], f_name:str)->None:
-        s3 = boto3.client('s3')
-        df = pd.DataFrame(news_output, columns=["post_time", "title", "content", "source", "link","keyword"])
-
-        # 데이터프레임을 parquet로 변환하여 메모리에서 처리
-        parquet_buffer = io.BytesIO()
-
-        # Pyspark에서 datetime이 깨지기 때문에 us 단위로 변환하여 저장
-        table = pa.Table.from_pandas(df=df)
-        pq.write_table(table, parquet_buffer, coerce_timestamps='us')
-        # df.to_parquet(parquet_buffer, engine="pyarrow")
-        parquet_buffer.seek(0)
-
-        try:
-            s3.upload_fileobj(Fileobj=parquet_buffer, Bucket=BUCKET_NAME, Key=f_name)
-            print(
-                f"[SBS] {len(df)} post are crawled.\n"
-                + f"[SBS] The data is successfully loaded at [{BUCKET_NAME}:{f_name}].\n"
-            )
-            s3.upload_fileobj(Fileobj=self._get_trigger_obj(), Bucket=BUCKET_NAME, Key=f"{MARKER_PATH}{self.NEWS_TAG}.json")
-        except Exception as e:
-            print("Error : {e}".format(e = e))
-
-    def _get_trigger_obj(self)->BytesIO:
-        json_data = {
-            "source": self.NEWS_TAG,
-            "start_time": self.request["start_time"].strftime("%Y-%m-%d %H:%M"),
-            "end_time": self.request["end_time"].strftime("%Y-%m-%d %H:%M"),
-            "status": "success"
-        }
-        json_bytes = json.dumps(json_data).encode("utf-8")
-        file_obj = io.BytesIO(json_bytes)
-        return file_obj
-    
-    def run(self):
-        search_result = self.get_search_result()
-        f_name = f'data/news/{self.request["start_time"]}_{self.request["end_time"]}/{self.request["keyword"]}_{self.NEWS_TAG}.parquet'
-        self.upload_s3(search_result, f_name)
-
-        # 로컬용
-        # self.save_to_parquet(search_result, f_name)
-
-
-# if __name__ == "__main__":
-#     if len(sys.argv) == 4:
-#         keyword = sys.argv[1]
-#         start_time_str = sys.argv[2]
-#         end_time_str = sys.argv[3]
-#         start_time = datetime.strptime(start_time_str, "%Y-%m-%dT%H:%M")
-#         end_time = datetime.strptime(end_time_str, "%Y-%m-%dT%H:%M")
-
-#         news_input : NewsRequest = {
-#             "keyword": keyword,
-#             "start_time": start_time,
-#             "end_time": end_time
-#         }
-#         sbs_crawler = SBSCrawler()
-#         search_result = sbs_crawler.get_search_result(news_input)
-#         sbs_crawler.save_to_parquet(search_result)
-#     else:
-#         print("Error : Missing Arguments | python sbs_crawler.py <keyword> <start_time> <end_time>", file=sys.stderr)
